chore: :tada: I Set Up Initial Project Configuration and Added Styles and Assets
I completed the initial setup of the project.
First, I installed the required packages using the `npm i` command.
Then, I added the necessary SVG files to the `public` folder.
I also created a `fonts` folder in `src/assets` and placed the "Vazir" font files inside it.
Finally, I added the `index.css` and `fonts.css` files in the `src/styles` folder and wrote the general styles for the project.