feat: :sparkles: I Added OTP API Integration and Handled SendOtpForm Submission
I created a new service in `src/services/auth.js` to handle the API request for sending OTP.
The `sendOTP` function uses the Axios instance from `api.js` to send a POST request to the `auth/send-otp` endpoint with the user's mobile number.
I also updated the `SendOtpForm` component to import and use this service.
In the `submitHandler`, after validating that the mobile number is 11 digits long, I called the `sendOTP` function. If the request is successful, I used `setStep(2)` to proceed to the next step.
If there’s an error, I logged the error message to the console.