feat: :sparkles: I Added Cookie Management for Tokens and Integrated It in CheckOtpForm
I created a new utility function `setCookie` in `src/utils/cookie.js` to manage storing tokens in the browser’s cookies.
This function sets the `accessToken` with a 1-day expiration and the `refreshToken` with a 30-day expiration.
I then updated the `CheckOtpForm` component to use this function.
After successfully verifying the OTP in the `submitHandler`, I called `setCookie` to store the tokens from the `response.data`.
If there's an error, I log the error message to the console.