feat: :sparkles: I Added User Profile Service and Integrated It with React Query in Router
I created a `user.js` file in `src/services/` that contains the `getProfile` function, which uses the Axios instance to make a `GET` request to the `user/whoami` endpoint.
Then, I integrated this service in the `Router.jsx` component by using React Query’s `useQuery` hook to fetch the user profile.
The query is identified by the key `["profile"]`, and it fetches the profile data when the component is rendered.
The returned `data` and `isLoading` states are used to manage the profile data and loading state within the component.