feat: :sparkles: I Added Token Refresh Mechanism in API Response Interceptor
I created a new `token.js` file in `src/services/` to handle the refresh token logic.
The `getNewToken` function retrieves the `refreshToken` from the cookies using `getCookie` and makes a request to the `auth/check-refresh-token` endpoint to get a new access token.
Then, I updated the Axios instance in `api.js` by adding a response interceptor.
When a `401 Unauthorized` error occurs, the interceptor checks if it's the first retry for the original request.
If so, it attempts to get a new access token using the refresh token and updates the cookies with the new tokens.
After successfully retrieving the new token, it retries the original request with the updated token.