feat: :sparkles: I Resolved Undefined Bug in User Profile Retrieval and Added Navigation After OTP Verification
I fixed the bug in `user.js` by modifying the `getProfile` function to handle undefined responses.
The function now returns either the response or `false`, ensuring that an undefined result is not passed along.
Additionally, I updated the `CheckOtpForm.jsx` component to include user redirection and identity fetching after OTP verification.
Upon successful verification of the OTP, the user is redirected to the home page using `useNavigate`. The `setCookie` function is used to store the tokens, and the `refetch` function from React Query is called to refresh the user profile data after login.