feat: :sparkles: I Added CategoryForm for Admin to Create New Categories
I developed the `CategoryForm` component, allowing admin users to create new categories by filling out fields for the name, slug, and icon.
The form uses `useState` to handle form input changes, storing the form data in a local state.
Additionally, I integrated this form into the `AdminPage` component, ensuring that it is accessible and usable within the admin panel.
This feature facilitates the creation of new categories in the application’s admin dashboard.