feat: :sparkles: I added react-hot-toast notifications for success and error handling in various forms.
I installed and integrated the react-hot-toast library to enhance user experience with notifications.
In the AddPost, CategoryForm, CheckOtpForm, and SendOtpForm components, I implemented toast notifications to display success or error messages based on form submission results.
This ensures that users receive real-time feedback on their actions, improving the overall usability and responsiveness of the application.